refactor: 괄호 디렉토리를 언더 스코어 디렉토리로 변경
