+++ app/(auth)/sign-up/_graphql/index.ts
