+++ app/(auth)/sign-up/_graphql/mutation.graphql
