+++ app/(tabs)/posts/_graphql/index.ts
