+++ app/(tabs)/posts/_graphql/query.generated.ts
