+++ app/posts/[postId]/_graphql/index.ts
