+++ app/posts/[postId]/_graphql/query.graphql
