+++ app/posts/new/_graphql/mutation.generated.ts
