+++ app/users/[accountId]/_graphql/query.generated.ts
